requirements.txt
google
googlesearch-python
